Fix: Implement persistent media playback notification
This commit resolves the issue of the media playback notification disappearing when the app is backgrounded or when switching tabs.

Key changes:
- Reverted the multi-WebView architecture to a single-WebView approach to simplify the codebase and fix lifecycle-related bugs.
- Integrated ExoPlayer into `MediaForegroundService` to handle background media playback. This is the standard and most robust way to manage background audio on Android.
- Implemented a handoff mechanism between the `WebView` and the `ExoPlayer` service. When the app is backgrounded or a tab is switched, the media state (URL and position) is transferred to the service, which resumes playback seamlessly.
- Introduced an `AndroidMediaBridge` and a `MediaState` data class to facilitate communication and state synchronization between the JavaScript in the `WebView` and the native Android code.
- Enhanced the injected JavaScript to reliably detect and monitor media playback, reporting state changes back to the app.

This new architecture ensures a continuous and smooth media playback experience for the user, with a persistent notification, similar to what is found in other major browsers.

+++ app/src/main/java/com/omymaxz/download/Tab.kt
@@ -14,7 +14,7 @@ data class Tab(
     var lastAccessTime: Long = System.currentTimeMillis(),
     var hasActiveMedia: Boolean = false,
     var isMediaPaused: Boolean = true,
-    var mediaPosition: Float = 0f,
+    var mediaPosition: Double = 0.0,
     var mediaUrl: String? = null,
     var mediaTitle: String? = null,
     // Add content caching
